requirements.txt
pygame==2.5.2
neat-python==0.92
numpy==1.24.3
Pillow==10.0.0 